Update First_Iteration_Hangman.py
This is my first attempt at making a hangman game in Python. There are probably ways I can optimize sections of the code, but I am happy with how it turned out. In the future, I might try to add a system which randomly selects a word to have the user guess, as apposed to having only one word which can only be changed by editing the code.