Create 21.py
This is a program I made for fun, which plays the children's game "21". In the game, you and another player take it in turns to say numbers from one to three, and as you go add the numbers towards a total. The moment a player makes the total go over 21, they lose. This program plays the game with the user, and plays it in two different ways depending on which mode you select, either  playing it perfectly, or randomly.